make everyViewmodel

+++ app/src/main/java/com/gymxy/gymxyone/domain/models/EachExerciseReps.kt
@@ -4,6 +4,6 @@ package com.gymxy.gymxyone.domain.models
  * weight in grams
  */
 data class EachExerciseReps (
-    var weight :Int = 0,
+    var weight :Long = 0,
     var reps : Int = 0
 )
